Welpe/requirements.txt
Mezzanine==4.2.2
python 2.7
django-gravatar
xlwt